Morse encoder for Console (!!) mode
Some requested encoders.. so.. here's an example. Might add more in the future.
Payload gets encoded with morse code and gets decoded when ran, revealing the malware (saved as: morse_payload.py).

This is a custom encoder. Doesn't need to be morse code but decided morse looks like fun. Morse is also well known.

Doesn't work with GUI mode because of the BASE64 image encoding. I 'might' fix this later.